Refactor other files to work with new refactor
