+++ src/quill_parser/src/file.rs
@@ -1,4 +1,8 @@
-use crate::{data_types::{DataP, EnumP}, definition::DefinitionP, identifier::IdentifierP};
+use crate::{
+    data_types::{DataP, EnumP},
+    definition::DefinitionP,
+    identifier::IdentifierP,
+};
 
 /// A single `.ql` file may export data types and definitions.
 /// This `File` struct contains the parsed abstract syntax tree of a file.
